added dedicated test class for AwsClientProviderImplTest.
